- Get rid of some compilation warnings in interactive.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@11495 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/interactive/thread.cpp
@@ -182,7 +182,7 @@ bool Semaphore::Post()
   int sem_val;
   bool success = false;
 
-  if(sem_getvalue(&semaphore_handle, &sem_val) == 0 && sem_val < impl->max_count)
+  if(sem_getvalue(&semaphore_handle, &sem_val) == 0 && sem_val < (int)impl->max_count)
   {
          success = (sem_post(&semaphore_handle) == 0);
   }
@@ -204,7 +204,7 @@ bool Semaphore::Post(unsigned count)
   }
 
   bool success = true;
-  for(int i = 0; i < count; ++i)
+  for(unsigned i = 0; i < count; ++i)
   {
          if(!(sem_val + i < impl->max_count && sem_post(&semaphore_handle) == 0))
          {
